Add files via upload
This is a basic mobile game that can be ported onto an android device. The basic premise is that you play a square, and reach checkpoints vs enemy AIs (need to do a bit more work on the AIs). Sound assets were left out intentionally. Created using the CodeNameOne cross-platform app development framework for Java.